更新验证码缓存，采用spring管理

+++ passport-main-service/src/main/java/com/sogou/upd/passport/common/utils/SMSUtil.java
@@ -15,11 +15,11 @@ public class SMSUtil {
 	
 	public static final long SEND_SMS_INTERVAL = 1000*60; // 发送短信验证码的间隔，1分钟只能发1条短信，单位ms
 	
-	public static final int MAX_SMS_COUNT_ONEDAY = 20; // 每日最多发送20短信验证码
+	public static final long MAX_SMS_COUNT_ONEDAY = 20; // 每日最多发送20短信验证码
 	
-	public static final int SMS_VALID = 30 * 60; // 短信验证码的有效期，30分钟
+	public static final long SMS_VALID = 30 * 60; // 短信验证码的有效期，30分钟
 
-	public static final int SMS_ONEDAY = 24 * 60 * 60; // 每日最多发送20短信验证码，时间 一天 1440分钟 ,单位s
+	public static final long SMS_ONEDAY = 24 * 60 * 60; // 每日最多发送20短信验证码，时间 一天 1440分钟 ,单位s
 
 	static final Logger logger = LoggerFactory.getLogger(SMSUtil.class);
